Work around references to cout/cerr in the object files
Closes #40.
